fix: Cast input values to int32 cuz apparently you can't add uint64 with int32
wtf

+++ input_system.m
@@ -5,7 +5,7 @@ function [car_num, seed, prng_selection] = get_user_input()
     printf('Welcome to Queue Simulator\n');
 
     % Get number of cars
-    car_num = cast(input('Enter no. of cars: '), 'uint64');
+    car_num = cast(input('Enter no. of cars: '), 'int32');
 
     % Get seed or randomize
     seed_decision = getline('Do you want to set your own seed value? [Y/n] : ');
@@ -27,7 +27,7 @@ function [car_num, seed, prng_selection] = get_user_input()
     disp('(1) Linear Congruential Generator (LCG)');
     disp('(2) Permuted Congruential Generator (PCG)');
     disp('(3) XOR-Shift-Reduced Plus (xorshiftr+)');
-    prng_input = uint64(input('Choose between 1 and 3: '));
+    prng_input = int32(input('Choose between 1 and 3: '));
 
     prng_selection = -1;
     switch(prng_input)
